更改机构

+++ qm-all/qm-biz/src/main/java/com/qmdj/dao/user/UserDAO.java
@@ -10,7 +10,7 @@ public interface UserDAO {
 	
 	public int saveUser(UserDO userDO);
 	
-	public UserDO queryUserByUserId(int userId);
+	public UserDO queryUserByUserId(long userId);
 	
 	public List<UserDO> queryUserList();
